Move notifcation-helper back to content-tooltip.
